# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AlertFilter": {
        "enum": [
          "open",
          "resolved",
          "all"
        ],
        "title": "AlertFilter",
        "type": "string"
      },
      "AlertOut": {
        "properties": {
          "acknowledged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "acknowledged_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged By"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "opened_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opened At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "severity": {
            "enum": [
              "info",
              "warning",
              "critical"
            ],
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AlertStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "severity",
          "status",
          "device_id",
          "title",
          "details",
          "opened_at",
          "resolved_at",
          "acknowledged_by",
          "acknowledged_at"
        ],
        "title": "AlertOut",
        "type": "object"
      },
      "AlertStatus": {
        "enum": [
          "pending",
          "open",
          "resolved"
        ],
        "title": "AlertStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InboxOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationOut"
            },
            "title": "Items",
            "type": "array"
          },
          "unread": {
            "title": "Unread",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "unread"
        ],
        "title": "InboxOut",
        "type": "object"
      },
...
      "NotificationOut": {
        "properties": {
          "alert_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Id"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event": {
            "title": "Event",
            "type": "string"
          },
          "home_id": {
            "format": "uuid",
            "title": "Home Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "severity": {
            "enum": [
              "info",
              "warning",
              "critical"
            ],
            "title": "Severity",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "home_id",
          "alert_id",
          "event",
          "severity",
          "title",
          "body",
          "created_at",
          "read_at"
        ],
        "title": "NotificationOut",
        "type": "object"
      },
...
      "PreferencesIn": {
        "properties": {
          "email_enabled": {
            "default": true,
            "title": "Email Enabled",
            "type": "boolean"
          },
          "min_email_severity": {
            "default": "warning",
            "enum": [
              "info",
              "warning",
              "critical"
            ],
            "title": "Min Email Severity",
            "type": "string"
          },
          "quiet_end": {
            "anyOf": [
              {
                "maximum": 23.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet End"
          },
          "quiet_start": {
            "anyOf": [
              {
                "maximum": 23.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Hour, home time",
            "title": "Quiet Start"
          }
        },
        "title": "PreferencesIn",
        "type": "object"
      },
      "PreferencesOut": {
        "properties": {
          "email_enabled": {
            "default": true,
            "title": "Email Enabled",
            "type": "boolean"
          },
          "min_email_severity": {
            "default": "warning",
            "enum": [
              "info",
              "warning",
              "critical"
            ],
            "title": "Min Email Severity",
            "type": "string"
          },
          "quiet_end": {
            "anyOf": [
              {
                "maximum": 23.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quiet End"
          },
          "quiet_start": {
            "anyOf": [
              {
                "maximum": 23.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Hour, home time",
            "title": "Quiet Start"
          }
        },
        "title": "PreferencesOut",
        "type": "object"
      },
      "QueuedOut": {
        "properties": {
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          }
        },
        "required": [
          "idempotency_key"
        ],
        "title": "QueuedOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebhookIn": {
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "rotate_secret": {
            "default": false,
            "title": "Rotate Secret",
            "type": "boolean"
          },
          "url": {
            "examples": [
              "https://example.com/hooks/smarthome"
            ],
            "maxLength": 500,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "WebhookIn",
        "type": "object"
      },
      "WebhookOut": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Only in the response that created or rotated it. Store it: it signs every delivery (X-Smarthome-Signature) and is never shown again.",
            "title": "Secret"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "title": "Updated By",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "enabled",
          "updated_by",
          "updated_at"
        ],
        "title": "WebhookOut",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Smart Home Hub API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/homes/{home_id}/alerts": {
      "get": {
        "description": "Alerts that have opened (newest first). A guest sees those of their devices.",
        "operationId": "list_alerts_homes__home_id__alerts_get",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AlertFilter",
              "default": "open"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertOut"
                  },
                  "title": "Response List Alerts Homes  Home Id  Alerts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Alerts",
        "tags": [
          "notifications"
        ]
      }
    },
    "/homes/{home_id}/alerts/{alert_id}/acknowledge": {
      "post": {
        "description": "Say you are on it. The alert stays open until whatever caused it clears.",
        "operationId": "acknowledge_homes__home_id__alerts__alert_id__acknowledge_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Alert Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Acknowledge",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/homes/{home_id}/notification-preferences": {
      "get": {
        "description": "Your own preferences for this home (defaults until you change them).",
        "operationId": "get_preferences_homes__home_id__notification_preferences_get",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Preferences",
        "tags": [
          "notifications"
        ]
      },
      "put": {
        "operationId": "put_preferences_homes__home_id__notification_preferences_put",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Preferences",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/homes/{home_id}/webhook": {
      "delete": {
        "operationId": "delete_webhook_homes__home_id__webhook_delete",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Webhook",
        "tags": [
          "notifications"
        ]
      },
      "get": {
        "operationId": "get_webhook_homes__home_id__webhook_get",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No webhook"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Webhook",
        "tags": [
          "notifications"
        ]
      },
      "put": {
        "description": "Where to POST alert transitions for this home. The response that creates the\nwebhook (or rotates its secret) is the only one that includes the secret.",
        "operationId": "put_webhook_homes__home_id__webhook_put",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Webhook",
        "tags": [
          "notifications"
        ]
      }
    },
    "/homes/{home_id}/webhook/test": {
      "post": {
        "description": "Queue a signed `ping` delivery to the webhook.",
        "operationId": "test_webhook_homes__home_id__webhook_test_post",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Webhook",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/me/notifications": {
      "get": {
        "description": "Your notifications across your homes, newest first. Page with `before` (the\n`created_at` of the last item you have).",
        "operationId": "inbox_me_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "unread_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unread Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboxOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inbox",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/me/notifications/{notification_id}/read": {
      "post": {
        "operationId": "mark_read_me_notifications__notification_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notification Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Read",
        "tags": [
          "notifications"
        ]
      }
    },
...
  }
}